requirements.txt
Flask
Pillow
onnxruntime
numpy
matplotlib
